Unify Background Database Observer logic
